// src/Game/Player.d.ts
import ECSManager from "../Engine/ECS/ECSManager";
import Vec3 from "../Engine/Maths/Vec3";
import Rendering from "../Engine/Rendering/Rendering";
import Scene from "../Engine/Rendering/Scene";
export default class Player {
    private scene;
    private rendering;
    private ecsManager;
    private positionComp;
    private movComp;
    private newspapers;
    private rotation;
    private game;
    private throwStrength;
    private throwPos;
    private throwVel;
    private throwCooldown;
    private throwTimer;
    private particleComp;
    private particleSpawner;
    private prepThrow;
    private throwRelease;
    constructor(scene: Scene, rendering: Rendering, ecsManager: ECSManager);
    init(): Promise<void>;
    respawn(): void;
    getPosition(): Vec3;
    getVelocity(): Vec3;
    throwPaper(): void;
    update(dt: number): void;
}